Allow overriding operator predicates via optional mapping arg (#9)
* Allow overriding operator predicates via optional mapping arg

The base library provides a base set of operators which are widely
supported across a range of SQL dialects. However, in some cases, users
may wish to override a particular operator for a more idiomatic/optimised
alternative for a given database. An example of this could be postgres
users preferring to use `= ANY` over `IN`.

* Bump version number to 0.2.0

Signed-off-by: Sam Lock <sam@swlock.co.uk>